More renames, remove dead code, general cleanup.
