+++ SongBrowserPlugin/UI/SongSortButton.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using SongBrowserPlugin.DataAccess;
 using UnityEngine.UI;
 
 namespace SongBrowserPlugin.UI
